+++ misc/dts-compare/src/remult3/remult3.d.ts
@@ -262,7 +262,7 @@ export interface LiveQuery<entityType> {
 export interface LiveQueryChangeInfo<entityType> {
     items: entityType[];
     changes: LiveQueryChange[];
-    applyChanges(prevState: entityType[]): entityType[];
+    applyChanges(prevState: entityType[] | undefined): entityType[];
 }
 export interface FindOptions<entityType> extends FindOptionsBase<entityType> {
     /** Determines the number of rows returned by the request, on the browser the default is 100 rows